--- src/pages/master/Master.tsx
@@ -1,12 +1,0 @@
-import MainContainer from "../../components/MainContainer";
-import MasterMain from "../../components/master/MasterMain";
-
-function Master() {
-  return (
-    <MainContainer>
-      <MasterMain />
-    </MainContainer>
-  );
-}
-
-export default Master;
